requirements.txt
anyio==3.6.2
certifi==2022.12.7
h11==0.14.0
httpcore==0.16.3
httpx==0.23.1
idna==3.4
numpy==1.24.1
opencv-python==4.7.0.68
pycodestyle==2.10.0
pydub==0.25.1
python-dotenv==0.21.0
python-telegram-bot==20.0b0
rfc3986==1.5.0
sniffio==1.3.0
tomli==2.0.1
yapf==0.32.0
